+++ src/client/java/com/duelco/JimmyToolsClient.java
@@ -4,7 +4,7 @@ import com.duelco._enum.Screen;
 import com.duelco.config.ModConfig;
 import com.duelco.handlers.BagHandler;
 import com.duelco.handlers.TransformationHelperHandler;
-import com.duelco.managers.BingoListener;
+import com.duelco.listeners.BingoListener;
 import com.duelco.managers.DataManager;
 import com.duelco.ui.screen.ScreenHandler;
 import net.fabricmc.api.ClientModInitializer;
