+++ src/client/java/com/duelco/managers/BingoManager.java
@@ -2,7 +2,7 @@ package com.duelco.managers;
 
 import com.duelco._enum.Screen;
 import com.duelco.config.ModConfig;
-import com.duelco.obj.BingoCard;
+import com.duelco.obj.bingo.BingoCard;
 import com.duelco.ui.screen.ScreenHandler;
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.text.Text;
