+++ src/client/java/com/duelco/ui/managers/BingoCardUIManager.java
@@ -1,8 +1,7 @@
 package com.duelco.ui.managers;
 
-import com.duelco.JimmyToolsClient;
 import com.duelco.config.ModConfig;
-import com.duelco.obj.BingoCard;
+import com.duelco.obj.bingo.BingoCard;
 import io.wispforest.owo.ui.component.Components;
 import io.wispforest.owo.ui.container.Containers;
 import io.wispforest.owo.ui.container.GridLayout;
