+++ src/client/java/com/duelco/ui/screen/BingoScreen.java
@@ -4,9 +4,8 @@ import com.duelco._enum.Screen;
 import com.duelco.handlers.ImageSelection;
 import com.duelco.handlers.ScreenCaptureHandler;
 import com.duelco.managers.BingoManager;
-import com.duelco.managers.DataManager;
 import com.duelco.managers.ToastManager;
-import com.duelco.obj.BingoCard;
+import com.duelco.obj.bingo.BingoCard;
 import com.duelco.ui.managers.BingoCardUIManager;
 import com.duelco.ui.managers.BingoMarkerUIManager;
 import io.wispforest.owo.ui.base.BaseOwoScreen;
